Created module for use aggregate
