--- src/api-authentication/src/main/java/org/rbernalop/apiauthentication/domain/exception/InvalidUserDataException.java
@@ -1,7 +1,0 @@
-package org.rbernalop.apiauthentication.domain.exception;
-
-public class InvalidUserDataException extends RuntimeException {
-    public InvalidUserDataException(String message) {
-        super(message);
-    }
-}
