--- src/api-authentication/src/main/java/org/rbernalop/apiauthentication/domain/exception/UserAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package org.rbernalop.apiauthentication.domain.exception;
-
-public class UserAlreadyExistsException extends RuntimeException {
-    public UserAlreadyExistsException(String message) {
-        super(message);
-    }
-}
