--- src/api-authentication/src/main/java/org/rbernalop/apiauthentication/domain/value_object/UserId.java
@@ -1,15 +1,0 @@
-package org.rbernalop.apiauthentication.domain.value_object;
-
-import jakarta.persistence.Embeddable;
-import lombok.EqualsAndHashCode;
-import lombok.NoArgsConstructor;
-import org.rbernalop.shared.domain.valueobject.Identifier;
-
-@Embeddable
-@NoArgsConstructor
-@EqualsAndHashCode
-public class UserId extends Identifier {
-    public UserId(String value) {
-        super(value);
-    }
-}
