--- src/api-authentication/src/main/java/org/rbernalop/apiauthentication/domain/value_object/UserName.java
@@ -1,22 +1,0 @@
-package org.rbernalop.apiauthentication.domain.value_object;
-
-
-import lombok.NoArgsConstructor;
-import org.rbernalop.apiauthentication.domain.exception.InvalidUserDataException;
-
-@NoArgsConstructor
-public class UserName {
-    private String name;
-
-    public UserName(String name) {
-        if(name == null || name.isBlank()) {
-            throw new InvalidUserDataException("User name cannot be empty");
-        }
-        this.name = name;
-    }
-
-    
-    public String getValue() {
-        return name;
-    }
-}
