--- src/api-authentication/src/main/java/org/rbernalop/apiauthentication/domain/value_object/UserPassword.java
@@ -1,17 +1,0 @@
-package org.rbernalop.apiauthentication.domain.value_object;
-
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class UserPassword {
-    private String password;
-
-    public UserPassword(String value) {
-        this.password = value;
-    }
-
-    
-    public String getValue() {
-        return password;
-    }
-}
