--- src/api-authentication/src/main/java/org/rbernalop/apiauthentication/infrastructure/controller/CustomExceptionHandler.java
@@ -1,49 +1,0 @@
-package org.rbernalop.apiauthentication.infrastructure.controller;
-
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
-import org.rbernalop.apiauthentication.domain.exception.InvalidUserDataException;
-import org.rbernalop.apiauthentication.domain.exception.UserAlreadyExistsException;
-import org.rbernalop.shared.domain.InvalidIdException;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.time.LocalDateTime;
-
-@ControllerAdvice
-@RestController
-public class CustomExceptionHandler extends ResponseEntityExceptionHandler {
-    private CustomError generateError(String message, HttpStatus httpStatus) {
-        return CustomError.builder()
-                .message(message)
-                .httpCode(httpStatus.value())
-                .timestamp(LocalDateTime.now())
-                .build();
-    }
-
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler({InvalidUserDataException.class, InvalidIdException.class})
-    public CustomError handleInvalidUserDataExceptionException(Exception ex) {
-        return generateError(ex.getCause().getMessage(), HttpStatus.BAD_REQUEST);
-    }
-
-    @ResponseStatus(HttpStatus.CONFLICT)
-    @ExceptionHandler(UserAlreadyExistsException.class)
-    public CustomError handleUserAlreadyExistsException(Exception ex) {
-        return generateError(ex.getCause().getMessage(), HttpStatus.CONFLICT);
-    }
-}
-
-@Builder
-@Getter
-@Setter
-class CustomError {
-    private String message;
-    private int httpCode;
-    private LocalDateTime timestamp;
-}
