+++ src/api-authentication/src/main/java/org/rbernalop/apiauthentication/user/domain/exception/InvalidUserDataException.java
@@ -1,0 +1,7 @@
+package org.rbernalop.apiauthentication.user.domain.exception;
+
+public class InvalidUserDataException extends RuntimeException {
+    public InvalidUserDataException(String message) {
+        super(message);
+    }
+}
