+++ src/api-authentication/src/main/java/org/rbernalop/apiauthentication/user/domain/exception/UserAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package org.rbernalop.apiauthentication.user.domain.exception;
+
+public class UserAlreadyExistsException extends RuntimeException {
+    public UserAlreadyExistsException(String message) {
+        super(message);
+    }
+}
