+++ src/api-authentication/src/main/java/org/rbernalop/apiauthentication/user/domain/value_object/UserPassword.java
@@ -1,0 +1,17 @@
+package org.rbernalop.apiauthentication.user.domain.value_object;
+
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+public class UserPassword {
+    private String password;
+
+    public UserPassword(String value) {
+        this.password = value;
+    }
+
+    
+    public String getValue() {
+        return password;
+    }
+}
