+++ src/api-authentication/src/main/java/org/rbernalop/apiauthentication/user/domain/value_object/UserSurname.java
@@ -1,0 +1,22 @@
+package org.rbernalop.apiauthentication.user.domain.value_object;
+
+import lombok.NoArgsConstructor;
+import org.rbernalop.apiauthentication.user.domain.exception.InvalidUserDataException;
+
+@NoArgsConstructor
+public class UserSurname {
+    private String surname;
+
+    public UserSurname(String surname) {
+        if(surname == null || surname.isBlank()) {
+            throw new InvalidUserDataException("User name cannot be empty");
+        }
+        this.surname = surname;
+    }
+
+    
+    public String getValue() {
+        return surname;
+    }
+}
+
